+++ build.ps1
@@ -183,8 +183,8 @@ process
             }
         }
 
-        # Loading Build Tasks defined in the .build/ folder (will override the ones imported above if same task name)
-        Get-ChildItem -Path ".build/" -Recurse -Include *.ps1 -ErrorAction Ignore | ForEach-Object {
+        # Loading Build Tasks defined in the source/tasks folder (will override the ones imported above if same task name)
+        Get-ChildItem -Path "source/tasks/" -Recurse -Include *.ps1 -ErrorAction Ignore | ForEach-Object {
             "Importing file $($_.BaseName)" | Write-Verbose
             . $_.FullName
         }