+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2020-06-22
+
 ### Added
 
 - Added cmdlet `Publish-WikiContent` that publishes the Wiki content